test: split energy model test
